rss-article-url-feeder/requirements.txt
nats-py==2.1.4
requests~=2.21.0
lxml~=4.6.3
